Move deletion test to FingerprintTest
