--- test/src/test/java/jenkins/fingerprints/FileFingerprintStorageTest.java
@@ -1,60 +1,0 @@
-/*
- * The MIT License
- *
- * Copyright (c) 2020, Sumit Sarin and Jenkins project contributors
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package jenkins.fingerprints;
-
-import hudson.Util;
-import hudson.model.Fingerprint;
-import org.junit.Rule;
-import org.junit.Test;
-import org.jvnet.hudson.test.JenkinsRule;
-
-import java.io.IOException;
-
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.nullValue;
-import static org.hamcrest.core.Is.is;
-import static org.hamcrest.core.IsNot.not;
-
-public class FileFingerprintStorageTest {
-
-    @Rule
-    public JenkinsRule rule = new JenkinsRule();
-
-    @Test
-    public void shouldDeleteFingerprint() throws IOException {
-        String id = Util.getDigestOf("shouldDeleteFingerprint");
-        new Fingerprint(null, "foo.jar", Util.fromHexString(id));
-
-        Fingerprint fingerprintLoaded = Fingerprint.load(id);
-        assertThat(fingerprintLoaded, is(not(nullValue())));
-
-        Fingerprint.delete(id);
-        fingerprintLoaded = Fingerprint.load(id);
-        assertThat(fingerprintLoaded, is(nullValue()));
-
-        Fingerprint.delete(id);
-        fingerprintLoaded = Fingerprint.load(id);
-        assertThat(fingerprintLoaded, is(nullValue()));
-    }
-}
